# server/common/swagger/Api.yaml
swagger: '2.0'
info:
  description: Api specs for ICDating
  version: 1.0.0-b1
  title: IC Dating API
  contact:
    email: charlie@aworkingapp.com
basePath: /api/v1
...
schemes:
  - http
paths:
...
  /account/hide/initiator/{initiatorAccountId}/receiver/{receiverAccountId}:
    post:
      tags:
        - discovery
      summary: Hide the receiver from the initiator.
      operationId: hideUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - name: initiatorAccountId
        in: path
        description: The user who perform the action of hiding someone in their discovery list
        required: true
        type: string
      - name: receiverAccountId
        in: path
        description: The user who received the hiding action and being hidden from the initiator
        required: true
        type: string
      responses:
        '201':
          description: successful operation
...